Consolidate TCP-specific common code into their own files.
The TCP keepalive tuning code has been moved into the newly added
files; this also allows future TCP-specific code to be added into
these files, without bloating the IP level code and establishes a
known file structure for other IP-based transports.

Remember: this is a no-op change, hence no API or functionality
was changed as part of this commit.
