+++ src/tcp.hpp
@@ -19,22 +19,19 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef __ZMQ_IP_HPP_INCLUDED__
-#define __ZMQ_IP_HPP_INCLUDED__
+#ifndef __ZMQ_TCP_HPP_INCLUDED__
+#define __ZMQ_TCP_HPP_INCLUDED__
 
 #include "fd.hpp"
 
 namespace zmq
 {
 
-    //  Same as socket(2), but allows for transparent tweaking the options.
-    fd_t open_socket (int domain_, int type_, int protocol_);
+    //  Tunes the supplied TCP socket for the best latency.
+    void tune_tcp_socket (fd_t s_);
 
-    //  Sets the socket into non-blocking mode.
-    void unblock_socket (fd_t s_);
-
-    //  Enable IPv4-mapping of addresses in case it is disabled by default.
-    void enable_ipv4_mapping (fd_t s_);
+    //  Tunes TCP keep-alives
+    void tune_tcp_keepalives (fd_t s_, int keepalive_, int keepalive_cnt_, int keepalive_idle_, int keepalive_intvl_);
 
 }
