+++ src/tcp_connecter.cpp
@@ -29,6 +29,7 @@
 #include "random.hpp"
 #include "err.hpp"
 #include "ip.hpp"
+#include "tcp.hpp"
 #include "address.hpp"
 #include "tcp_address.hpp"
 #include "session_base.hpp"
